backend/fastapi/pyproject.toml
[project]
name = "YTP-hackathon-backend"
version = "0.1.0"
description = "backend"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "lucasw0908" }, 
    { name = "IanWen103026" },
    { name = "jx06T" },
    { name = "Watermelon-1234"}
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.49",
]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
    "ruff>=0.13.0",
]

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings --cov=app"
markers = [
    "webtest: mark a test as a web test",
]

[tool.fastapi]
entrypoint = "main:app"
